dashboards/streamlit-analytics-dashboard/requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
supabase==2.3.4
python-dotenv==1.0.0
openpyxl==3.1.2
reportlab==4.0.9
numpy==1.26.3
